+++ src/main/java/cart/application/OrderService.java
@@ -1,6 +1,5 @@
 package cart.application;
 
-import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -14,7 +13,6 @@ import cart.dao.MemberCouponDao;
 import cart.dao.OrderDao;
 import cart.dao.OrderItemDao;
 import cart.domain.CartItem;
-import cart.domain.Coupon;
 import cart.domain.Member;
 import cart.domain.MemberCoupon;
 import cart.domain.Money;
@@ -22,10 +20,7 @@ import cart.domain.Order;
 import cart.domain.OrderItem;
 import cart.domain.PriceCalculator;
 import cart.dto.request.CartItemRequest;
-import cart.dto.request.MemberCouponAddRequest;
 import cart.dto.request.OrderRequest;
-import cart.dto.response.CouponResponse;
-import cart.dto.response.CouponsResponse;
 import cart.dto.response.MemberCouponResponse;
 import cart.dto.response.MemberCouponsResponse;
 import cart.dto.response.OrderDetailResponse;
@@ -157,42 +152,5 @@ public class OrderService {
         orderDao.validate(member.getId(), orderId);
         orderItemDao.deleteAllOf(orderId);
         orderDao.delete(orderId);
-    }
-
-    public MemberCouponsResponse findMemberCoupons(Member member, List<Long> cartItemIds) {
-        List<CartItem> cartItems = cartItemDao.findByIds(cartItemIds);
-        Money price = PriceCalculator.calculate(cartItems);
-        List<MemberCoupon> memberCoupons = memberCouponDao.findByMemberId(member.getId());
-
-        return new MemberCouponsResponse(
-            memberCoupons.stream()
-                .map(memberCoupon -> MemberCouponResponse.of(memberCoupon, price))
-                .collect(Collectors.toUnmodifiableList()));
-    }
-
-    public CouponsResponse findAllCoupons() {
-        List<Coupon> coupons = couponDao.findAll();
-        return new CouponsResponse(coupons.stream().map(CouponResponse::of).collect(Collectors.toUnmodifiableList()));
-    }
-
-    @Transactional
-    public Long addMemberCoupon(Member member, Long couponId, MemberCouponAddRequest memberCouponAddRequest) {
-        Coupon coupon = couponDao.findById(couponId);
-        checkExistence(coupon);
-
-        MemberCoupon memberCoupon = new MemberCoupon(
-            null,
-            member,
-            coupon,
-            Timestamp.valueOf(memberCouponAddRequest.getExpiredAt())
-        );
-
-        return memberCouponDao.save(memberCoupon);
-    }
-
-    private void checkExistence(Coupon coupon) {
-        if (coupon == null) {
-            throw new BadRequestException(ExceptionType.COUPON_NO_EXIST);
-        }
     }
 }
